Add plugin to create and clear running notes in code
Whenever I read code, I like to write little comments that explain what
I've understood so far. I created the laindoc plugin to do this - it
creates a new comment that is prefixed with laindoc, as well as clears
them when the clear command is ran to quickly remove them.

I didn't want to bother reading nvim documentation to do this so Claude
was very useful at getting this done quickly.